WIP Corrected view frustum for all viewers
My initial `adjustFrustum`, used by `LightedCameraControls`, was only available in
classic, since it was attached to the `CameraContext` (formerly `RotationContext`).
Similarly, the `recordCamera` behavior necessary for 2D and some 3D exporters was
also unavailable in all contexts.  Now I've moved those to the `WorkerStateContext`,
and I've restored `RotationContext` as it was.

Some refactoring and renaming is indicated.  I should split out the viewer-only behavior
of `WorkerStateContext` as `ViewerContext`.  I may also want to do a viewer-editor
split on the worker side, perhaps.
