+++ online/src/app/classic/classic.jsx
@@ -9,7 +9,7 @@ import { SceneEditor } from './components/editor.jsx';
 import { useWorkerClient } from "../../workerClient/index.js";
 import { OrbitsDialog } from "./dialogs/orbits.jsx";
 import { ShapesDialog } from "./dialogs/shapes.jsx";
-import { CameraProvider } from "../../viewer/solid/cameracontext.jsx";
+import { RotationProvider } from "../../viewer/solid/rotation.jsx";
 import { PolytopesDialog } from "./dialogs/polytopes.jsx";
 import { ErrorAlert } from "./components/alert.jsx";
 
@@ -23,7 +23,7 @@ export const ClassicEditor = ( props ) =>
 
   let alertRoot;
   return (
-    <CameraProvider>
+    <RotationProvider>
     <div id='classic' ref={alertRoot} style={{ display: 'grid', 'grid-template-rows': '1fr' }} class='whitesmoke-bkgd'>
       <div id='editor-main' class='grid-cols-1-min whitesmoke-bkgd' >
 
@@ -50,7 +50,7 @@ export const ClassicEditor = ( props ) =>
       </div>
       <ErrorAlert/>
     </div>
-    </CameraProvider>
+    </RotationProvider>
   )
 }
